requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
whisperx @ git+https://github.com/m-bain/whisperx.git
faster-whisper==1.2.1
openai>=1.50.0
python-dotenv>=1.0.1
pandas==2.2.3
tqdm>=4.66.4
orjson>=3.10.7
protobuf==4.25.3
pydantic>=2.8.2
sentence-transformers==2.2.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
transformers==4.37.2
huggingface-hub==0.36.0
tokenizers==0.15.2
numpy>=2.0.0
pyannote.audio==3.4.0
scikit-learn>=1.5.1
tabulate>=0.9.0
langcodes>=3.4.0
streamlit>=1.37.0
matplotlib>=3.9.0
